// ui-catalogue/src/components/molecules/RegulatedNonFinancialSystem/types.tsx
interface Ii18nRegulatedNonFinancialSystem {
  'RegulatedNonFinancialSystem:title': string
  'RegulatedNonFinancialSystem:subtitle': string
  'RegulatedNonFinancialSystem:flagImpagos': string
  'RegulatedNonFinancialSystem:flagDeudaAtrasada': string
  'RegulatedNonFinancialSystem:flagDeudasImpagadasInfocorp': string
  'RegulatedNonFinancialSystem:flagInformacionNoDisponible': string
  'RegulatedNonFinancialSystem:diasAtraso': string
}

export interface IRegulatedNonFinancialSystemRequestForm {
  regulatedPeriodo1: string
  regulatedPeriodo2: string
  regulatedPeriodo3: string
  regulatedPeriodo4: string
  regulatedPeriodo5: string
  regulatedPeriodo6: string
  regulatedFlagImpagos1: string
  regulatedFlagImpagos2: string
  regulatedFlagImpagos3: string
  regulatedFlagImpagos4: string
  regulatedFlagImpagos5: string
  regulatedFlagImpagos6: string
  regulatedFlagDeudaAtrasada1: string
  regulatedFlagDeudaAtrasada2: string
  regulatedFlagDeudaAtrasada3: string
  regulatedFlagDeudaAtrasada4: string
  regulatedFlagDeudaAtrasada5: string
  regulatedFlagDeudaAtrasada6: string
  regulatedFlagDeudasImpagadasInfocorp1: string
  regulatedFlagDeudasImpagadasInfocorp2: string
  regulatedFlagDeudasImpagadasInfocorp3: string
  regulatedFlagDeudasImpagadasInfocorp4: string
  regulatedFlagDeudasImpagadasInfocorp5: string
  regulatedFlagDeudasImpagadasInfocorp6: string
  regulatedFlagInformacionNoDisponible1: string
  regulatedFlagInformacionNoDisponible2: string
  regulatedFlagInformacionNoDisponible3: string
  regulatedFlagInformacionNoDisponible4: string
  regulatedFlagInformacionNoDisponible5: string
  regulatedFlagInformacionNoDisponible6: string
  regulatedDiasAtraso1: string
  regulatedDiasAtraso2: string
  regulatedDiasAtraso3: string
  regulatedDiasAtraso4: string
  regulatedDiasAtraso5: string
  regulatedDiasAtraso6: string
  nonRegulatedPeriodo1: string
  nonRegulatedPeriodo2: string
  nonRegulatedPeriodo3: string
  nonRegulatedPeriodo4: string
  nonRegulatedPeriodo5: string
  nonRegulatedPeriodo6: string
  nonRegulatedFlagImpagos1: string
  nonRegulatedFlagImpagos2: string
  nonRegulatedFlagImpagos3: string
  nonRegulatedFlagImpagos4: string
  nonRegulatedFlagImpagos5: string
  nonRegulatedFlagImpagos6: string
  nonRegulatedFlagDeudaAtrasada1: string
  nonRegulatedFlagDeudaAtrasada2: string
  nonRegulatedFlagDeudaAtrasada3: string
  nonRegulatedFlagDeudaAtrasada4: string
  nonRegulatedFlagDeudaAtrasada5: string
  nonRegulatedFlagDeudaAtrasada6: string
  nonRegulatedFlagDeudasImpagadasInfocorp1: string
  nonRegulatedFlagDeudasImpagadasInfocorp2: string
  nonRegulatedFlagDeudasImpagadasInfocorp3: string
  nonRegulatedFlagDeudasImpagadasInfocorp4: string
  nonRegulatedFlagDeudasImpagadasInfocorp5: string
  nonRegulatedFlagDeudasImpagadasInfocorp6: string
  nonRegulatedFlagInformacionNoDisponible1: string
  nonRegulatedFlagInformacionNoDisponible2: string
  nonRegulatedFlagInformacionNoDisponible3: string
  nonRegulatedFlagInformacionNoDisponible4: string
  nonRegulatedFlagInformacionNoDisponible5: string
  nonRegulatedFlagInformacionNoDisponible6: string
  nonRegulatedDiasAtraso1: string
  nonRegulatedDiasAtraso2: string
  nonRegulatedDiasAtraso3: string
  nonRegulatedDiasAtraso4: string
  nonRegulatedDiasAtraso5: string
  nonRegulatedDiasAtraso6: string
}

export interface IRegulatedNonFinancialSystem {
  i18n: Ii18nRegulatedNonFinancialSystem
  data: IRegulatedNonFinancialSystemRequestForm
}